Moved ScopeTests to ReferenceTests
