+++ Source/ReferenceTests/ReferenceTests.csproj
@@ -55,6 +55,7 @@
     <Compile Include="Language\FunctionTests.cs" />
     <Compile Include="Language\InvocationTests.cs" />
     <Compile Include="Language\Operators\RangeOperatorTests_7_4.cs" />
+    <Compile Include="Language\ScopeTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReferenceHost.cs" />
     <Compile Include="Language\PipelineExecutionTests.cs" />
@@ -125,7 +126,5 @@
   <Target Name="AfterBuild">
   </Target>
   -->
-  <ItemGroup>
-    <Folder Include="Parsing\" />
-  </ItemGroup>
+  <ItemGroup />
 </Project>
